requirements.txt
Django==1.10.1
pytest-django==3.0.0
mock==2.0.0
